src/backend/requirements.txt

fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2
pydantic==2.5.3
python-dotenv==1.0.0
